fix(surfaces): Correct definition of closed surface
The previous definition required vertex neighborhoods to be cycles
of length 3 or more. This incorrectly classified the tetrahedral
graph (K4), a 3-simplex, as a closed surface.

This commit updates the check to require neighborhood cycles to have
a length of at least 4, which correctly identifies minimal
triangulations of a sphere like the octahedral graph.